Rearranged utils and feeder file
